Begin complete overhaul of Incomplete streaming handling
An incomplete parsing result should not always be transformed into a failure.

For example,the streaming versions of take_while() and its variants will return Incomplete
if the last character still meets the conditions. However, if that is the last byte of data,
we should be able to accept that as a parse result.

This commit reworks the Err type to split Incomplete into IncompleteFail and IncompleteSuccess.
The next steps are porting the streaming stuffs to use the new variants.